+++ src/libponyc/codegen/gencall.h
@@ -27,7 +27,7 @@ LLVMValueRef gencall_alloc(compile_t* c, reach_type_t* t);
 
 LLVMValueRef gencall_allocstruct(compile_t* c, reach_type_t* t);
 
-void gencall_throw(compile_t* c);
+void gencall_error(compile_t* c);
 
 void gencall_memcpy(compile_t* c, LLVMValueRef dst, LLVMValueRef src,
   LLVMValueRef n);