+++ src/libponyrt/gc/cycle.h
@@ -18,6 +18,8 @@ void ponyint_cycle_unblock(pony_ctx_t* ctx, pony_actor_t* actor);
 
 void ponyint_cycle_ack(pony_ctx_t* ctx, size_t token);
 
+void ponyint_cycle_check_blocked(pony_ctx_t* ctx);
+
 void ponyint_cycle_terminate(pony_ctx_t* ctx);
 
 bool ponyint_is_cycle(pony_actor_t* actor);